requirements.txt
asyncio==3.4.3
playwright==1.47.0
pyMuPDF==1.24.10
fl
pyzbar==0.1.9
pillow==10.4.0
opencv-python==4.10.0.84
pandas==2.2.3
ruff==0.6.6